Add decline attendance button to attendance confirm page (#131)
* Move components into Admission folder

* Add button, fix some type issues

* Enable cannot attend button

* Add confirmation popup

* Update padding and margins
